Bugfix
Each label and each image are separate results. That way we might get multiple labels (and images) for the same incoming / outgoing node. That can lead to not getting all the results.
The fix only corrected this for the labels. For the images, we might still returning multiple results. I could correct for it, but since we don't know, wether first image will be found, I raterh decided to leave this in.